+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Added constants, constructor, and additional conversions to `Version`
 - Changed `patch` field to default to `0` in `Version` instead of using `Option`
 - Reworked byte and string array types in Vulkan structs
+- Removed dummy `EMPTY` bitflags used in empty `bitflags!` instances
 
 ### Added
 - Added `Debug` requirement to extension struct traits (e.g., `ExtendsDeviceCreateInfo`)